Struct cup_perps::prop_perps::CupPerp

pub struct CupPerp { /* private fields */ }

Implementations

impl CupPerp

pub fn instantiate_pair( pair: String, leverage: Decimal, funding_coeff: Decimal, exch: Decimal, deposit: Bucket ) -> ComponentAddress

Instantiates the Component

The initial deposit’s lp is not issued, to prevent division by zero, but the initial deposit may be worth close to 0 followed by the actual liqidity providers

Arguments

pair - String shown in LP token metadata

leverage - The leverage of both cups’ exposure when at equilibrium

funding_coeff - Scales the funding rebate (funding * cup size ratio)

exch - Initial pair exchange rate

deposit - Liquidity to be locked in the pair forever

// resim 0.8.0
> resim call-function [package] CupPerp instantiate_pair "5x BTC/USD" 5 0.75 28345 10,[xrd]
 
// wallet-sdk 0.6.0-beta.11
.withdrawFromAccountByAmount(
    [account], 
    100, 
    [xrd address])
.takeFromWorktop(
    [xrd address],
    "buck")
.callFunction(
    [package], 
    "CupPerp", 
    "instantiate_pair", 
    [String("RAND/USD"), Decimal(1000), Bucket("buck")])
.callMethod(
     [account], 
     "deposit_batch", 
     [Expression("ENTIRE_WORKTOP")])
.build()
.toString();

pub fn update(&mut self)

Updates the system state after any price change

// scrypto 0.8.0
> resim call-method [component] update
 
// wallet-sdk 0.6.0-beta.11
.callMethod(
    [component],
    "update",
    [])
.callMethod(
     [account], 
     "deposit_batch", 
     [Expression("ENTIRE_WORKTOP")])
.build()
.toString();

pub fn deposit(&mut self, side: bool, funds: Bucket) -> Bucket

Deposits the stable asset to the select cup

Arguments

side - Which cup to deposit into? Long is true, Short is false.

funds - Funds to be deposited

// resim 0.8.0
> resim call-method [component] deposit true 10,[xrd]
 
// wallet-sdk 0.6.0-beta.11
.withdrawFromAccountByAmount(
    [account],
    10,
    [xrd address])
.takeFromWorktop(
    [xrd address],
    "buck1")
.callMethod(
    [component],
    "deposit",
    [Bool(false), Bucket("buck1")])
.callMethod(
     [account], 
     "deposit_batch", 
     [Expression("ENTIRE_WORKTOP")])
.build()
.toString();

pub fn withdraw(&mut self, funds: Bucket) -> Bucket

Withdraws the liquidity with an LP token

// scrypto 0.8.0
> resim call-method [component] withdraw 1337,[ShortLp]
 
// wallet-sdk 0.6.0-beta.11
.withdrawFromAccountByAmount(
    [account],
    420,
    [LongLp address])
.takeFromWorktop(
    [LongLp address],
    "buck1")
.callMethod(
    [component],
    "withdraw",
    [Bucket("buck1")])
.callMethod(
     [account], 
     "deposit_batch", 
     [Expression("ENTIRE_WORKTOP")])
.build()
.toString();

pub fn set_oracle(&mut self, n: Decimal)

Sets the built-in oracle to a given number

Arguments

n - Decimal exchange rate

// scrypto 0.8.0
> resim call-method [component] set_oracle 4
 
// wallet-sdk 0.6.0-beta.11
.callMethod(
    [component],
    "set_oracle",
    [Decimal(5138008)])
.callMethod(
     [account], 
     "deposit_batch", 
     [Expression("ENTIRE_WORKTOP")])
.build()
.toString();

pub fn show_cups(&self) -> (Decimal, Decimal)

Shows the current amount of tokens in both cups

// scrypto 0.8.0
> resim call-method [component] get_lp_resource_addr 

pub fn get_lp_resource_addr(&self) -> (ResourceAddress, ResourceAddress)

Returns the ResourceAddress of both Lp tokens

// scrypto 0.8.0
> resim call-method [component] get_lp_resource_addr 

pub fn value(&self, long: Decimal, short: Decimal) -> Decimal

Returns the total value of a given quantity of Lp tokens

Arguments

long - Amount of LongLp tokens

short - Amount of ShortLp tokens

// scrypto 0.8.0
> resim call-method [component] value 7 0